refactor: 태그 타입 외부 파일로 분리리
